Buff Alignment - Multiple Timeline Edge Case
Interesting edge case here in a multiple timeline scenario.
If Crit / DH stats differ and there is a Crit/DH relevant buff,
the potency of the hover for the timeline would be that of your current config, not that timeline's config
By saving a reference of the config into the Potency we can reference it at all times.