AST Potency not displayed
Improved: Healer dots without initial potency will now shot the buffs according to snapshot timing.

Problem: If the dot is refreshed before the dot tick, it will have neither initial potency nor dot potency, and so will not show the buffs that would've been applied to it had it been allowed to tick.